+++ src/cli.ts
@@ -1,14 +1,15 @@
 #!/usr/bin/env node
 
-import { parsedArgs } from "./lib/argumentParser";
-import { getVersion } from "./lib/console";
-const args = parsedArgs['_'];
+import argumentParser from './cli/argumentParser';
+import { getVersion } from './cli/console';
+
+const args = argumentParser._;
 
 if (args.includes('version')) {
-    getVersion();
-    process.exit(0);
-} 
-if (args.includes("help")) {
-    console.log("Help page");
-    process.exit(0);
+  getVersion();
+  process.exit(0);
+}
+if (args.includes('help')) {
+  console.log('Help page');
+  process.exit(0);
 }
