+++ src/lib/struct/facebookConfiguration.ts
@@ -1,0 +1,4 @@
+export interface FacebookConfiguration {
+  username: string;
+  password: string;
+}
