--- src/struct/facebookConfiguration.ts
@@ -1,4 +1,0 @@
-export interface FacebookConfiguration {
-    username: string;
-    password: string;
-}
